fixed borrow and issue responses

+++ library management/library-management/src/main/java/com/librarymanagement/service/BooksService.java
@@ -28,6 +28,8 @@ public class BooksService {
         return booksRepository.save(book);
     }
 
+
+
     public Books updateBook(Integer id, Books updatedBook) {
         Books book = booksRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid book Id:" + id));
@@ -40,10 +42,39 @@ public class BooksService {
     }
 
     public void deleteBook(Integer id) {
+        Books book = booksRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Invalid book Id:" + id));
+
+        boolean hasActiveBorrows = book.getBorrows()
+                .stream()
+                .anyMatch(b -> Boolean.FALSE.equals(b.getIsReturned()));
+
+        if (hasActiveBorrows) {
+            throw new IllegalStateException("Cannot delete book that is currently issued to a user");
+        }
+
         booksRepository.deleteById(id);
     }
 
 
+    public Books saveBookWithAuthorName(Books book, String authorName, AuthorsService authorsService) {
+        Optional<Authors> existingAuthor = authorsService.getAllAuthors()
+                .stream()
+                .filter(a -> a.getName().equalsIgnoreCase(authorName))
+                .findFirst();
+
+        Authors author;
+        if (existingAuthor.isPresent()) {
+            author = existingAuthor.get();
+        } else {
+            author = new Authors();
+            author.setName(authorName);
+            author = authorsService.saveAuthor(author);
+        }
+
+        book.setAuthorId(author.getAuthorId());
+        return booksRepository.save(book);
+    }
 }
